arrumando os arquivos dentro da pasta routes
